feat(xlang): add bfloat16 support (#3605)
## Why?

Add cross-language `bfloat16` support alongside the existing
reduced-precision `float16` support so reduced-precision scalars and
dense arrays can round-trip across runtimes with spec-backed behavior
and documentation.


## What does this PR do?

- adds `bfloat16` scalar and array/list carriers plus
serializer/type-resolution wiring in Java, C++, C#, Python, Go, and
Rust, with the generated/object-codec paths updated where needed
- extends xlang and schema-evolution handling for reduced-precision
fields, including array skip logic and field-ordering/type-id fixes
required for consistent cross-language behavior
- adds cross-language coverage for reduced-precision structs and updates
runtime tests across the touched implementations
- refreshes the xlang specification, type-mapping docs, and language
guides/READMEs to document `bfloat16` and reduced-precision array
support

## Related issues

#1017 
Closes #3286 
Closes #3283
Closes #3285 

## AI Contribution Checklist



- [ ] Substantial AI assistance was used in this PR: `yes` / `no`
- [ ] If `yes`, I included a completed [AI Contribution
Checklist](https://github.com/apache/fory/blob/main/AI_POLICY.md#9-contributor-checklist-for-ai-assisted-prs)
in this PR description and the required `AI Usage Disclosure`.
- [ ] If `yes`, my PR description includes the required `ai_review`
summary and screenshot evidence of the final clean AI review results
from both fresh reviewers on the current PR diff or current HEAD after
the latest code changes.



## Does this PR introduce any user-facing change?



- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

+++ java/fory-core/src/test/java/org/apache/fory/type/TypesTest.java
@@ -28,6 +28,6 @@ public class TypesTest {
   public void testGetClassForFloatTypeIds() {
     assertSame(Types.getClassForTypeId(Types.FLOAT8), Float.class);
     assertSame(Types.getClassForTypeId(Types.FLOAT16), Float16.class);
-    assertSame(Types.getClassForTypeId(Types.BFLOAT16), Float.class);
+    assertSame(Types.getClassForTypeId(Types.BFLOAT16), BFloat16.class);
   }
 }
